more numpy 1.20 fixes

+++ meshio/ply/_ply.py
@@ -22,8 +22,8 @@ ply_to_numpy_dtype = {
     # [u]char is often used as [u]int, e.g., from Wikipedia:
     # > The word 'list' indicates that the data is a list of values, the first of which
     # > is the number of entries in the list (represented as a 'uchar' in this case).
-    "char": numpy.int,
-    "uchar": numpy.uint,
+    "char": numpy.int8,
+    "uchar": numpy.uint8,
     "short": numpy.int16,
     "ushort": numpy.uint16,
     "int": numpy.int32,
